Move default to javascript-kernel, add interface
