+++ packages/javascript-kernel/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@jupyterlite/kernel",
+  "name": "@jupyterlite/javascript-kernel",
   "version": "0.1.0",
-  "description": "JupyterLite - Kernel",
+  "description": "JupyterLite - JavaScript Kernel",
   "homepage": "https://github.com/jtpio/jupyterlite",
   "bugs": {
     "url": "https://github.com/jtpio/jupyterlite/issues"
@@ -43,9 +43,7 @@
     "watch": "tsc -b --watch"
   },
   "dependencies": {
-    "@jupyterlab/services": "^6.0.5",
-    "@lumino/disposable": "^1.4.3",
-    "@lumino/signaling": "^1.4.3"
+    "@jupyterlite/kernel": "^0.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
