+++ packages/javascript-kernel/src/index.ts
@@ -1,0 +1,2 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
