+++ packages/javascript-kernel/test/jskernel.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+describe('JavaScript Kernel', () => {
+  describe('#constructor()', () => {
+    it('should create a Kernel instance', () => {
+      expect(true).toBe(true);
+    });
+  });
+});
