+++ packages/javascript-kernel/tsconfig.test.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfigbase.test",
+  "include": ["src/**/*", "test/**/*"],
+  "references": [
+    {
+      "path": "."
+    }
+  ]
+}
