+++ packages/kernel/src/index.ts
@@ -2,3 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 
 export * from './kernels';
+export * from './default';
+import * as Kernel from './kernel';
+
+export { Kernel };
